Added files via upload
Added the appointment page, added documentation to several files in HealthNet (including the HTML files), and fixed a typo in index where a button wasn't spelled properly. The background in appointment somewhat fits the color scheme in index, but it isn't perfect, nor does anything else link to appointment since we didn't have anything other than the sign up and sign in buttons, so I didn't know where we wanted to have the button to schedule an appointment.